Drill down into folders
Added 'open folder' buttons to all the folders; when you click them, that folder takes over the grid. You can't get back out of the folder and up the hierarchy yet, though.

The folders are listed at the tops of the lists now.

Also did some work on the layout with the columns scrolling horizontally